--- src/state/atoms/index.js
@@ -1,3 +1,0 @@
-// export individual atoms from here to simplify importing atoms in components
-export { headerTitle } from './headerTitle';
-export { userState } from './userState';
