+++ src/state/headerTitle/index.js
@@ -1,0 +1,2 @@
+// Atom exports
+export { headerTitle } from './atoms';
