--- src/state/selectors/currentUserState.js
@@ -1,11 +1,0 @@
-import { selector } from 'recoil';
-import { userState } from '../atoms';
-
-export const currentUserState = selector({
-  key: 'currentUserState',
-  get: ({ get }) => {
-    const currentUser = get(userState);
-
-    return currentUser;
-  },
-});
