--- src/state/selectors/index.js
@@ -1,1 +1,0 @@
-// export individual selectors from here to simplify importing selectors in components
